+++ cmd/forwarder/main.go
@@ -133,7 +133,7 @@ func main() {
 		fmt.Printf("could not create %s stream error: %v\n", natsStreamName, err)
 		return
 	}
-	log.Info(fmt.Sprintf("%s jetStream createdOrUpdated", natsStreamName))
+	log.Info(natsStreamName + " jetStream createdOrUpdated")
 
 	consumers, err := consumer.NewConsumers(
 		log,
@@ -168,11 +168,11 @@ func main() {
 	app.RegisterInfraRoutes(r)
 	app.RunHTTPServer(gCtx, g, cfg.AppConfig.Port, r)
 
-	log.Info(fmt.Sprintf(`Started %s`, cfg.AppConfig.Name))
+	log.Info("Started " + cfg.AppConfig.Name)
 
 	if err := g.Wait(); err != nil {
 		log.Error(err.Error())
 	}
 
-	log.Info(fmt.Sprintf(`Main done %s`, cfg.AppConfig.Name))
+	log.Info("Main done " + cfg.AppConfig.Name)
 }